chore(@clayui/multi-step-nav): update MultiStepNav component stories
